{libivh} TimeProfiling

+++ ivhd/libivhd/include/utils/Math.h
@@ -4,11 +4,11 @@
 
 namespace ivhd::math
 { 
-	static int randInt(int min, int max)
+	static size_t randInt(size_t min, size_t max)
 	{
 		std::random_device rd;
 		std::mt19937 mt(rd());
-		std::uniform_int_distribution<int> dist(min, max);
+		std::uniform_int_distribution<size_t> dist(min, max);
 
 		return dist(mt);
 	}
